+++ session_cpp/src/boundingbox.cpp
@@ -11,11 +11,11 @@
 namespace session_cpp {
 
 BoundingBox::BoundingBox() 
-    : center(0.0f, 0.0f, 0.0f),
-      x_axis(1.0f, 0.0f, 0.0f),
-      y_axis(0.0f, 1.0f, 0.0f),
-      z_axis(0.0f, 0.0f, 1.0f),
-      half_size(0.5f, 0.5f, 0.5f),
+    : center(0.0, 0.0, 0.0),
+      x_axis(1.0, 0.0, 0.0),
+      y_axis(0.0, 1.0, 0.0),
+      z_axis(0.0, 0.0, 1.0),
+      half_size(0.5, 0.5, 0.5),
       guid(::guid()),
       name("my_boundingbox") {}
 
@@ -28,31 +28,31 @@ BoundingBox::BoundingBox(const Point& center, const Vector& x_axis, const Vector
       guid(::guid()),
       name("my_boundingbox") {}
 
-BoundingBox::BoundingBox(const Plane& plane, float dx, float dy, float dz)
+BoundingBox::BoundingBox(const Plane& plane, double dx, double dy, double dz)
     : center(plane.origin()),
       x_axis(plane.x_axis()),
       y_axis(plane.y_axis()),
       z_axis(plane.z_axis()),
-      half_size(dx * 0.5f, dy * 0.5f, dz * 0.5f),
+      half_size(dx * 0.5, dy * 0.5, dz * 0.5),
       guid(::guid()),
       name("") {}
 
-BoundingBox BoundingBox::from_point(const Point& point, float inflate_amount) {
-    BoundingBox box(point, Vector(1.0f, 0.0f, 0.0f), Vector(0.0f, 1.0f, 0.0f), Vector(0.0f, 0.0f, 1.0f), Vector(inflate_amount, inflate_amount, inflate_amount));
+BoundingBox BoundingBox::from_point(const Point& point, double inflate_amount) {
+    BoundingBox box(point, Vector(1.0, 0.0, 0.0), Vector(0.0, 1.0, 0.0), Vector(0.0, 0.0, 1.0), Vector(inflate_amount, inflate_amount, inflate_amount));
     return box;
 }
 
-BoundingBox BoundingBox::from_points(const std::vector<Point>& points, float inflate_amount) {
+BoundingBox BoundingBox::from_points(const std::vector<Point>& points, double inflate_amount) {
     if (points.empty()) {
         return BoundingBox();
     }
     
-    float min_x = std::numeric_limits<float>::max();
-    float min_y = std::numeric_limits<float>::max();
-    float min_z = std::numeric_limits<float>::max();
-    float max_x = std::numeric_limits<float>::lowest();
-    float max_y = std::numeric_limits<float>::lowest();
-    float max_z = std::numeric_limits<float>::lowest();
+    double min_x = std::numeric_limits<double>::max();
+    double min_y = std::numeric_limits<double>::max();
+    double min_z = std::numeric_limits<double>::max();
+    double max_x = std::numeric_limits<double>::lowest();
+    double max_y = std::numeric_limits<double>::lowest();
+    double max_z = std::numeric_limits<double>::lowest();
     
     for (const auto& pt : points) {
         min_x = std::min(min_x, pt.x());
@@ -63,34 +63,34 @@ BoundingBox BoundingBox::from_points(const std::vector<Point>& points, float inf
         max_z = std::max(max_z, pt.z());
     }
     
-    Point center((min_x + max_x) * 0.5f, (min_y + max_y) * 0.5f, (min_z + max_z) * 0.5f);
+    Point center((min_x + max_x) * 0.5, (min_y + max_y) * 0.5, (min_z + max_z) * 0.5);
     Vector half_size(
-        (max_x - min_x) * 0.5f + inflate_amount,
-        (max_y - min_y) * 0.5f + inflate_amount,
-        (max_z - min_z) * 0.5f + inflate_amount
+        (max_x - min_x) * 0.5 + inflate_amount,
+        (max_y - min_y) * 0.5 + inflate_amount,
+        (max_z - min_z) * 0.5 + inflate_amount
     );
-    return BoundingBox(center, Vector(1.0f, 0.0f, 0.0f), Vector(0.0f, 1.0f, 0.0f), Vector(0.0f, 0.0f, 1.0f), half_size);
+    return BoundingBox(center, Vector(1.0, 0.0, 0.0), Vector(0.0, 1.0, 0.0), Vector(0.0, 0.0, 1.0), half_size);
 }
 
-BoundingBox BoundingBox::from_line(const Line& line, float inflate_amount) {
+BoundingBox BoundingBox::from_line(const Line& line, double inflate_amount) {
     std::vector<Point> points = {line.start(), line.end()};
     return from_points(points, inflate_amount);
 }
 
-BoundingBox BoundingBox::from_polyline(const Polyline& polyline, float inflate_amount) {
+BoundingBox BoundingBox::from_polyline(const Polyline& polyline, double inflate_amount) {
     return from_points(polyline.points, inflate_amount);
 }
 
-BoundingBox BoundingBox::from_mesh(const Mesh& mesh, float inflate_amount) {
+BoundingBox BoundingBox::from_mesh(const Mesh& mesh, double inflate_amount) {
     auto [vertices, faces] = mesh.to_vertices_and_faces();
     return from_points(vertices, inflate_amount);
 }
 
-BoundingBox BoundingBox::from_pointcloud(const PointCloud& pointcloud, float inflate_amount) {
+BoundingBox BoundingBox::from_pointcloud(const PointCloud& pointcloud, double inflate_amount) {
     return from_points(pointcloud.points, inflate_amount);
 }
 
-Point BoundingBox::point_at(float x, float y, float z) const {
+Point BoundingBox::point_at(double x, double y, double z) const {
     return Point(
         center.x() + x * x_axis.x() + y * y_axis.x() + z * z_axis.x(),
         center.y() + x * x_axis.y() + y * y_axis.y() + z * z_axis.y(),
@@ -146,7 +146,7 @@ std::array<Point, 10> BoundingBox::two_rectangles() const {
     return result;
 }
 
-void BoundingBox::inflate(float amount) {
+void BoundingBox::inflate(double amount) {
     half_size = Vector(
         half_size.x() + amount,
         half_size.y() + amount,
@@ -157,19 +157,19 @@ void BoundingBox::inflate(float amount) {
 bool BoundingBox::separating_plane_exists(const Vector& relative_position, const Vector& axis, const BoundingBox& box1, const BoundingBox& box2) {
     Vector rp = relative_position;
     Vector ax = axis;
-    float dot_rp = std::abs(rp.dot(ax));
+    double dot_rp = std::abs(rp.dot(ax));
     
     Vector v1 = box1.x_axis * box1.half_size.x();
     Vector v2 = box1.y_axis * box1.half_size.y();
     Vector v3 = box1.z_axis * box1.half_size.z();
     Vector ax1 = axis;
-    float proj1 = std::abs(v1.dot(ax1)) + std::abs(v2.dot(ax1)) + std::abs(v3.dot(ax1));
+    double proj1 = std::abs(v1.dot(ax1)) + std::abs(v2.dot(ax1)) + std::abs(v3.dot(ax1));
     
     Vector v4 = box2.x_axis * box2.half_size.x();
     Vector v5 = box2.y_axis * box2.half_size.y();
     Vector v6 = box2.z_axis * box2.half_size.z();
     Vector ax2 = axis;
-    float proj2 = std::abs(v4.dot(ax2)) + std::abs(v5.dot(ax2)) + std::abs(v6.dot(ax2));
+    double proj2 = std::abs(v4.dot(ax2)) + std::abs(v5.dot(ax2)) + std::abs(v6.dot(ax2));
     return dot_rp > (proj1 + proj2);
 }
